Experimental block cursor mode
When using a block cursor (like in vim mode), this would make the
brackets highlight when the cursor is at bracket char.  This is
different that outside adjacent which will highlight when the cursor is
outside on either side and different than normal mode which will
highlight only when the cursor is between.  This will highlight when
the char is right in front of the bracket or between, but won’t
highlight when the cursor is right after the closing bracket.